Mingw64 build fixes

+++ src/C4Game.cpp
@@ -1250,7 +1250,7 @@ C4Object* C4Game::FindObject(C4ID id,
 								// ActionTarget
 								if (!pActionTarget || (pActionDef && ((cObj->Action.Target==pActionTarget) || (cObj->Action.Target2==pActionTarget)) ))
 									// Container
-									if ( !pContainer || (cObj->Contained == pContainer) || ((reinterpret_cast<long>(pContainer)==NO_CONTAINER) && !cObj->Contained) || ((reinterpret_cast<long>(pContainer)==ANY_CONTAINER) && cObj->Contained) )
+									if ( !pContainer || (cObj->Contained == pContainer) || ((reinterpret_cast<intptr_t>(pContainer)==NO_CONTAINER) && !cObj->Contained) || ((reinterpret_cast<intptr_t>(pContainer)==ANY_CONTAINER) && cObj->Contained) )
 										// Owner
 										if ((iOwner==ANY_OWNER) || (cObj->Owner==iOwner))
 											// Area
