+++ src/platform/OpenURL.cpp
@@ -35,7 +35,7 @@
 bool OpenURL(const char *szURL)
 {
 #ifdef _WIN32
-	if ((int)ShellExecute(NULL, "open", szURL, NULL, NULL, SW_SHOW) > 32)
+	if ((intptr_t)ShellExecute(NULL, "open", szURL, NULL, NULL, SW_SHOW) > 32)
 		return true;
 #endif
 #ifdef WITH_GLIB
